Bootstrap the "pywgen" program
We define the base command line parser for the "pywgen" program using
argparse. This contains a --help flag (built-in) and a --version one,
which uses pkg_resources to compute the version from packaging metadata.

The program is defined as a console script entry point in
setuptools.setup(), which points to the main() function in pywgen
module. The program can then be run either simply as `pywgen` or using
`python3 -m pywgen`.

Along the way, we add a "mypy" tox environment to make sure our code
type checks.