Allow "numerals" to be None in generate_passwords()
Re-reading the documentation of "-n, --numerals" option in pwgen, it
seems that we may not either require or require not having a number in
generated password, but also need to allow possibly having one. This is
similar to the "capitalize" option. We handle this by allowing None as
"numerals" value, meaning that the generated password may have a number
but this is not required.