Declare "num_pw" postitional argument in parser
It has no default value, since it depends on whether the program is
executed in interactive mode or not. We document this in module
docstring, which is then used as program's description. In order to keep
the text properly formatted, we need to change the help formatter class
of argparse's parser so that it does not transform the docstring (which
is already properly formatted).