Implement logic for default value of "num_pw" argument
When the program is run interactively, 160 passwords are generated by
pwgen (20 lines of 8 values). We reproduce this logic at argument
parsing level to set the default value of "num_pw" (if not specified)
depending on whether stdout is a TTY (meaning it's not redirected, hence
in interactive mode).

We implement this by defining a custom Action class for the "num_pw"
argument of our parser.

In tests, we use pytest's "capsys" fixture to toggle stdout capture.