Introduce a pronounceable_choice() function
This will be used as a generator for phonemes to generate a
pronounceable password. It works by randomly choosing a phoneme value in
given set and yielding it if it forms a pronounceable syllabe with the
previously yielded value.

In tests, we replace secrets.choice() by next() in order to consume
values in a predictable way.