Add a generate_pronounceable_password() function
This function will be used instead of generate_password() function in
the program when asked for pronounceable password (the default in
"pwgen").

Since phonemes may consists of more than one character, we need a
special logic to control the length of generated password.

In tests, we replace pronounceable_choice() by an iterator over the
"choice" parameter to get predictable values generated. The test
illustrates how a password generated by processing the first needed
items of the choice may be rejected when it does not match conditions
specified by keywords (e.g. presence of a symbols).