Extract common password check function
Function to generate passwords have a common logic to check that
produced passwords respect presence of capital letters, numbers or
symbols. We extract this into a check_password() function.

By the way, as we add unit tests, we notice that the original code was
wrong as we were missing proper distinction of None and False values.
The consequence in program execution is that we might get a password
without a requested character class; for instance running "pywgen 2 10
-y" might produce one password without an symbol. Fixing this here.