Add support for -B/--ambiguous option
The idea is to extend "remove_chars" option when this flag is true at
parser level. Hence, we define a custom argparse.Action similar to
"append_const" action from argparse but appending to a string instead of
a list. Since we append to the existing "remove_chars" option, we do not
need to modify anything but the parser since generate_password (or
generate_pronounceable_password) will not receive the extra "ambiguous"
argument but an extended "remove_chars" one instead.

This makes pywgen on par with pwgen as far a characters set
configuration is concerned. Only the -v/--no-vowel is missing, but it
actually seems redundant with -B/--ambiguous.